Delete least used files, update UI for favorite page
