+++ core/src/main/java/com/minotawr/gamers/di/DatabaseModule.kt
@@ -2,7 +2,10 @@ package com.minotawr.gamers.di
 
 import android.content.Context
 import androidx.room.Room
+import com.minotawr.gamers.core.BuildConfig
 import com.minotawr.gamers.data.local.GamersDatabase
+import net.sqlcipher.database.SQLiteDatabase
+import net.sqlcipher.database.SupportFactory
 import org.koin.android.ext.koin.androidContext
 import org.koin.dsl.module
 
@@ -14,10 +17,14 @@ val databaseModule = module {
     single { get<GamersDatabase>().tagsDao() }
 }
 
-fun provideDatabase(context: Context) =
-    Room.databaseBuilder(
+fun provideDatabase(context: Context): GamersDatabase {
+    val passphrase: ByteArray = SQLiteDatabase.getBytes(BuildConfig.DB_PASSPHRASE.toCharArray())
+    val factory = SupportFactory(passphrase)
+    return Room.databaseBuilder(
         context.applicationContext,
         GamersDatabase::class.java,
         "Gamers.db"
     ).fallbackToDestructiveMigration()
+        .openHelperFactory(factory)
         .build()
+}
