+++ favoritegame/src/main/java/com/minotawr/favoritegame/ui/FavoriteAdapter.kt
@@ -3,40 +3,14 @@ package com.minotawr.favoritegame.ui
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
-import com.bumptech.glide.Glide
-import com.minotawr.favoritegame.databinding.AdapterFavoriteBinding
+import com.minotawr.gamers.databinding.AdapterMainBinding
 import com.minotawr.gamers.domain.model.Game
 
-class FavoriteAdapter: RecyclerView.Adapter<FavoriteAdapter.FavoriteViewHolder>() {
+class FavoriteAdapter : RecyclerView.Adapter<FavoriteViewHolder>() {
 
     private val list = mutableListOf<Game>()
 
     var delegate: FavoriteListDelegate? = null
-
-    class FavoriteViewHolder(private val binding: AdapterFavoriteBinding): RecyclerView.ViewHolder(binding.root) {
-        fun bind(item: Game) {
-            Glide.with(itemView.context)
-                .load(item.backgroundImage)
-                .into(binding.imgGame)
-
-            binding.tvTitle.text = item.name
-            binding.tvRating.text = String.format("Rated %.2f out of 5 stars", item.rating)
-
-            item.genres?.take(2)?.forEachIndexed { index, genre ->
-                if (index < 1) {
-                    binding.tvGenre1.text = genre.name
-                } else {
-                    binding.tvGenre2.text = genre.name
-                }
-            }
-
-            binding.imgGame.contentDescription = "Image of ${item.name}'s banner."
-        }
-    }
-
-    interface FavoriteListDelegate {
-        fun onItemClick(item: Game)
-    }
 
     fun setList(data: List<Game>) {
         val currentSize = list.size
@@ -48,7 +22,7 @@ class FavoriteAdapter: RecyclerView.Adapter<FavoriteAdapter.FavoriteViewHolder>(
     }
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): FavoriteViewHolder {
-        val binding = AdapterFavoriteBinding.inflate(LayoutInflater.from(parent.context), parent, false)
+        val binding = AdapterMainBinding.inflate(LayoutInflater.from(parent.context), parent, false)
         return FavoriteViewHolder(binding)
     }
