+++ favoritegame/src/main/java/com/minotawr/favoritegame/ui/FavoriteListDelegate.kt
@@ -1,0 +1,7 @@
+package com.minotawr.favoritegame.ui
+
+import com.minotawr.gamers.domain.model.Game
+
+interface FavoriteListDelegate {
+    fun onItemClick(item: Game)
+}
