+++ gradle.properties
@@ -26,3 +26,5 @@ version_name = 1.0.0
 
 API_URL = https://api.rawg.io/api/
 API_KEY = ebbdf1b129ea49f88661d7115b8b8378
+
+DB_PASSPHRASE = gamers
